Avoid build exceptions when generating Javadoc
Prior to this change, `./gradlew build` included Javadoc generation by
default and would fail raising a FileNotFoundException due to the
absence of docs/src/api/spring-javadoc.css. This file should be treated
as optional, and in any case its absence should not cause a build
failure.

Now the build script no longer attempts to apply buildSrc/docs.gradle to
the :docs subproject, but instead defines a simpler version of the
:docs:api task.  Instructions in the existing readme are unchanged; run
`./gradlew build` to compile/test/package the project, and run
`./gradlew :docs:api` to generate Javadoc.